extract components into separate files
